refactor(routes): Split product_helpers into specialized modules
Tarea 3 - Complete refactorization of product_helpers.py

Changes:
- Created app/routes/product_services/ package with 4 specialized modules:
  * adapters.py (~180 LOC, MI: 65.63): HTTP/HTTPX handling, backend adapters
  * barcode_lookup.py (~200 LOC, MI: 60.84): Barcode lookup and OpenFoodFacts integration
  * legacy_ocr_compat.py (~180 LOC, MI: 62.96): Legacy OCR compatibility layer
  * ocr_processor.py (~445 LOC, MI: 42.43): OCR scanning and processing
- Refactored product_helpers.py to 111 LOC (MI: 100)
  * Now serves as orchestration layer with imports and re-exports
  * Maintains backward compatibility with all tests
- Updated __init__.py for proper module exports

Results:
- All 1792 tests passing ✅
- All modules: MI > 40 ✅
- Improved code organization and maintainability
- Clear separation of concerns
- Easier to test and modify individual components

Metrics:
- Original: product_helpers.py 828 LOC, MI 16.67 (critical)
- After: 1115 LOC total, distributed across modules
- All modules have MI >= 40 (target achieved)
- Main file: 111 LOC (near original target of ~150)

References:
- plan/2025-12-27-product-helpers-refactoring.md (Tarea 3 plan)
- plan/2025-12-26-refactoring-architectural-plan.md (Master plan - Phase 1)
