+++ app/routes/product_helpers.py
@@ -1,23 +1,17 @@
-import logging
-import os
-import re
-import sys
-import tempfile
-import inspect
-from datetime import datetime
-from typing import Optional, Union
-from unittest.mock import Mock as _MockType
+"""
+Product Helpers - Main Route Module
 
-from fastapi import APIRouter, Depends, File, HTTPException, UploadFile, status
-from app.models.product import (
-    BarcodeRequest, ProductResponse, ErrorResponse,
-    ScanResponse, LowConfidenceScanResponse, Nutriments
-)
-from app.services.openfoodfacts import openfoodfacts_service
-from app.services.cache import cache_service
-from app.services import nutrition_ocr
-from app.services import cache as cache_module
-from app.services import openfoodfacts as openfoodfacts_module
+Orchestrates product operations by importing from specialized service modules.
+Maintains backward compatibility with existing tests and route registration.
+
+Task: Phase 1 Tarea 3 - Product Helpers Refactoring
+
+Module Structure:
+- product_services.ocr_processor: OCR scanning and processing
+- product_services.barcode_lookup: Barcode and OpenFoodFacts integration
+- product_services.legacy_ocr_compat: Legacy OCR compatibility
+- product_services.adapters: HTTP and backend adapters
+"""
 
 # Try to import aiofiles for backward compatibility with tests
 try:
@@ -25,7 +19,63 @@ try:
 except ImportError:
     aiofiles = None
 
+# ─────────────────────────────────────────────────────────────
+# Import from Specialized Modules
+# ─────────────────────────────────────────────────────────────
+
+# Legacy OCR Compatibility
+from .product_services.legacy_ocr_compat import (
+    _import_legacy_ocr,
+    _get_legacy_service,
+    _get_legacy_parser_callable,
+    _get_legacy_text_extractor,
+    _normalize_legacy_ocr_result,
+    _normalize_parsed_payload,
+    _parse_text_with_legacy_parser,
+    _run_legacy_text_pipeline,
+)
+
+# OCR Processor
+from .product_services.ocr_processor import (
+    _run_local_ocr,
+    extract_nutrients_from_image,
+    call_external_ocr,
+    scan_nutrition_label,
+    scan_label_with_external_ocr,
+)
+
+# Barcode Lookup
+from .product_services.barcode_lookup import (
+    get_product_by_barcode,
+)
+
+# Backend Adapters
+from .product_services.adapters import (
+    _get_cache_backend,
+    _get_openfoodfacts_backend,
+    _coerce_exception_class,
+    _get_httpx_attr,
+    _get_httpx_exceptions,
+    _raise_http_exception,
+    _is_http_exception,
+    _route_post,
+    _ensure_request_context,
+    _write_bytes_to_tempfile,
+)
+
+# Auth context re-exports for backward compatibility with tests
+from app.services.auth import RequestContext, get_optional_request_context
+
+# Database service re-export for backward compatibility with tests
+from app.services.database import db_service
+
+
+# ─────────────────────────────────────────────────────────────
+# Re-exports for Backward Compatibility
+# ─────────────────────────────────────────────────────────────
+
 __all__ = [
+    # Legacy OCR functions
     "_normalize_legacy_ocr_result",
     "_run_local_ocr",
     "extract_nutrients_from_image",
@@ -47,9 +97,11 @@ __all__ = [
     "_normalize_parsed_payload",
     "_parse_text_with_legacy_parser",
     "_run_legacy_text_pipeline",
+    # Main routes
     "get_product_by_barcode",
     "scan_nutrition_label",
     "scan_label_with_external_ocr",
+    # Utilities
     "aiofiles",
     # Auth context re-exports for backward compatibility with tests
     "RequestContext",
@@ -57,772 +109,3 @@ __all__ = [
     # Database service re-export for backward compatibility with tests
     "db_service",
 ]
-
-
-def _normalize_legacy_ocr_result(payload: dict, *, default_engine: str) -> Optional[dict]:
-    if not isinstance(payload, dict):
-        return None
-
-    normalized = dict(payload)
-    raw_text = normalized.get('raw_text') or normalized.get('text') or ''
-
-    parsed = normalized.get('parsed_nutriments') or normalized.get('nutrition_data') or normalized.get('nutrients')
-    if not isinstance(parsed, dict):
-        parsed = {}
-    normalized['parsed_nutriments'] = parsed
-    if not raw_text and parsed:
-        raw_text = 'legacy_ocr_output'
-    normalized['raw_text'] = raw_text
-    normalized.setdefault('nutrients', parsed)
-    normalized.setdefault('nutrition_data', parsed)
-
-    serving_size = normalized.get('serving_size')
-    serving_info = normalized.get('serving_info')
-    if not isinstance(serving_info, dict):
-        serving_info = {'detected': serving_size, 'unit': None}
-    else:
-        serving_info = serving_info.copy()
-    serving_info.setdefault('detected', serving_size)
-    serving_info.setdefault('unit', None)
-    normalized['serving_info'] = serving_info
-    normalized.setdefault('serving_size', serving_info.get('detected'))
-
-    try:
-        normalized['confidence'] = float(normalized.get('confidence', 0.0))
-    except (TypeError, ValueError):
-        normalized['confidence'] = 0.0
-
-    processing_details = normalized.get('processing_details')
-    if not isinstance(processing_details, dict):
-        processing_details = {}
-    processing_details.setdefault('ocr_engine', default_engine)
-    normalized['processing_details'] = processing_details
-
-    found = normalized.get('found_nutrients')
-    if not isinstance(found, list):
-        found = list(parsed.keys())
-    normalized['found_nutrients'] = found
-    missing = normalized.get('missing_required')
-    if not isinstance(missing, list):
-        missing = []
-    normalized['missing_required'] = missing
-
-    return normalized
-
-
-async def _run_local_ocr(image_path: str, *, debug: bool = False) -> dict:
-    proxied_result = extract_nutrients_from_image(image_path, debug=debug)
-    if inspect.isawaitable(proxied_result):
-        proxied_result = await proxied_result
-    if proxied_result:
-        return proxied_result
-
-    legacy_ocr = _import_legacy_ocr()
-    if legacy_ocr is not None:
-        normalized = await _run_legacy_text_pipeline(legacy_ocr, image_path)
-        if normalized:
-            return normalized
-
-        legacy_service = _get_legacy_service(legacy_ocr)
-        legacy_callable = getattr(legacy_service, 'extract_nutrients', None) if legacy_service else None
-        if legacy_callable:
-            legacy_output = legacy_callable(image_path)
-            if inspect.isawaitable(legacy_output):
-                legacy_output = await legacy_output
-            normalized = _normalize_legacy_ocr_result(legacy_output or {}, default_engine='legacy_ocr') if legacy_output else None
-            if normalized:
-                return normalized
-
-    return proxied_result
-
-
-def extract_nutrients_from_image(*args, **kwargs):
-    """Proxy to nutrition_ocr.extract_nutrients_from_image to ease patching in tests."""
-    return nutrition_ocr.extract_nutrients_from_image(*args, **kwargs)
-
-
-def call_external_ocr(*args, **kwargs):
-    """Proxy to nutrition_ocr.call_external_ocr to ease patching in tests."""
-    return nutrition_ocr.call_external_ocr(*args, **kwargs)
-def _get_cache_backend():
-    return getattr(cache_module, 'cache_service', cache_service)
-
-
-def _get_openfoodfacts_backend():
-    return getattr(openfoodfacts_module, 'openfoodfacts_service', openfoodfacts_service)
-from app.services.database import db_service
-from app.repositories.product_repository import ProductRepository
-from app.models.product import Product
-from app.services.analytics_service import AnalyticsService
-from app.services.auth import RequestContext, get_optional_request_context
-
-# Task: Phase 2 Batch 6 - Analytics Service Extraction
-analytics_service = AnalyticsService(db_service)
-import aiofiles
-
-
-class _TimeoutProxy(Exception):
-    """Sentinel used when httpx.TimeoutException is unavailable."""
-
-
-class _RequestErrorProxy(Exception):
-    """Sentinel used when httpx.RequestError is unavailable."""
-
-
-try:
-    import httpx as _httpx_module
-except Exception:  # pragma: no cover - triggered when httpx missing
-    _httpx_module = None
-
-if isinstance(_httpx_module, _MockType) or _httpx_module is None:
-    httpx = _httpx_module
-    _HTTPX_TIMEOUT_CLASS = _TimeoutProxy
-    _HTTPX_REQUEST_ERROR_CLASS = _RequestErrorProxy
-else:
-    httpx = _httpx_module
-    _HTTPX_TIMEOUT_CLASS = getattr(httpx, 'TimeoutException', _TimeoutProxy)
-    _HTTPX_REQUEST_ERROR_CLASS = getattr(httpx, 'RequestError', _RequestErrorProxy)
-
-
-logger = logging.getLogger(__name__)
-router = APIRouter()
-
-
-def _coerce_exception_class(candidate):
-    if isinstance(candidate, _MockType):
-        return Exception
-    if isinstance(candidate, type) and issubclass(candidate, BaseException):
-        return candidate
-    return Exception
-
-
-def _get_httpx_attr(httpx_module, attr_name: str, fallback):
-    if isinstance(httpx_module, _MockType) and attr_name not in httpx_module.__dict__:
-        return fallback
-    return getattr(httpx_module, attr_name, fallback)
-
-
-def _get_httpx_exceptions():
-    """Return the active Timeout and RequestError classes honoring patched modules."""
-    httpx_module = sys.modules.get('httpx', httpx)
-    timeout_exc = _coerce_exception_class(
-        _get_httpx_attr(
-            httpx_module,
-            'TimeoutException',
-            _HTTPX_TIMEOUT_CLASS
-        )
-    )
-    request_error_exc = _coerce_exception_class(
-        _get_httpx_attr(
-            httpx_module,
-            'RequestError',
-            _HTTPX_REQUEST_ERROR_CLASS
-        )
-    )
-    return timeout_exc, request_error_exc
-
-
-def _raise_http_exception(status_code: int, detail: str):
-    """Create HTTPException, honoring FastAPI test doubles when present."""
-    fastapi_module = sys.modules.get('fastapi')
-    exc_cls = getattr(fastapi_module, 'HTTPException', HTTPException)
-    return exc_cls(status_code=status_code, detail=detail)
-
-
-def _is_http_exception(exc: Exception) -> bool:
-    """Determine whether an exception is FastAPI's HTTPException or a patched variant."""
-    fastapi_module = sys.modules.get('fastapi')
-    exc_cls = getattr(fastapi_module, 'HTTPException', HTTPException)
-    return isinstance(exc, exc_cls)
-
-
-def _route_post(*args, **kwargs):
-    decorator = router.post(*args, **kwargs)
-    if _MockType is not None and isinstance(decorator, _MockType):
-        def _noop(func):
-            return func
-        return _noop
-    return decorator
-
-
-def _ensure_request_context(context: RequestContext) -> RequestContext:
-    """Provide a safe fallback when route functions are invoked directly in tests."""
-    if isinstance(context, RequestContext):
-        return context
-    # Any other type (including FastAPI Depends placeholders or mocks) -> anonymous context
-    return RequestContext(user=None, session_id=None, token=None)
-
-
-async def _write_bytes_to_tempfile(path: str, data: bytes):
-    """Write upload bytes to disk, tolerating aiofiles mocks used in tests."""
-    open_callable = getattr(aiofiles, 'open', None)
-    if open_callable is None:
-        with open(path, 'wb') as f:  # pragma: no cover - safety fallback
-            f.write(data)
-        return
-
-    ctx = open_callable(path, 'wb')
-    if inspect.isawaitable(ctx):
-        ctx = await ctx
-
-    if hasattr(ctx, '__aenter__'):
-        async with ctx as f:
-            write_result = f.write(data)
-            if inspect.isawaitable(write_result):
-                await write_result
-        return
-
-    # As a last resort, fall back to synchronous write.
-    with open(path, 'wb') as f:  # pragma: no cover - safety fallback
-        f.write(data)
-
-
-def _import_legacy_ocr():
-    try:
-        from app.services import ocr as legacy_ocr
-        return legacy_ocr
-    except ImportError:
-        return None
-
-
-def _get_legacy_service(legacy_ocr):
-    return getattr(legacy_ocr, 'ocr_service', None) if legacy_ocr else None
-
-
-def _get_legacy_parser_callable(legacy_ocr):
-    parser_module = getattr(legacy_ocr, 'nutrition_parser', None) if legacy_ocr else None
-    return getattr(parser_module, 'parse_nutrition_text', None) if parser_module else None
-
-
-def _get_legacy_text_extractor(legacy_ocr):
-    service = _get_legacy_service(legacy_ocr)
-    return getattr(service, 'extract_text', None) if service else None
-
-
-async def _normalize_parsed_payload(raw_text: str, parsed_payload: Optional[dict], *, engine_label: str) -> Optional[dict]:
-    parsed_payload = parsed_payload or {}
-    payload = {
-        'raw_text': raw_text or '',
-        'parsed_nutriments': parsed_payload.get('nutrition_data') or parsed_payload.get('parsed_nutriments') or {},
-        'serving_size': parsed_payload.get('serving_size'),
-        'confidence': parsed_payload.get('confidence', 0.0),
-        'processing_details': parsed_payload.get('processing_details') or {}
-    }
-    return _normalize_legacy_ocr_result(payload, default_engine=engine_label)
-
-
-async def _parse_text_with_legacy_parser(legacy_ocr, raw_text: str, *, engine_label: str) -> Optional[dict]:
-    parser_callable = _get_legacy_parser_callable(legacy_ocr)
-    if parser_callable is None:
-        return None
-    parsed = parser_callable(raw_text or '')
-    if inspect.isawaitable(parsed):
-        parsed = await parsed
-    if not isinstance(parsed, dict):
-        parsed = {}
-    return await _normalize_parsed_payload(raw_text, parsed, engine_label=engine_label)
-
-
-async def _run_legacy_text_pipeline(legacy_ocr, image_path: str) -> Optional[dict]:
-    extractor = _get_legacy_text_extractor(legacy_ocr)
-    if extractor is None:
-        return None
-    raw_text = extractor(image_path)
-    if inspect.isawaitable(raw_text):
-        raw_text = await raw_text
-    return await _parse_text_with_legacy_parser(legacy_ocr, raw_text or '', engine_label='legacy_ocr')
-
-
-async def _normalize_external_payload(result, legacy_ocr, *, engine_label: str) -> Optional[dict]:
-    if not result:
-        return None
-    if isinstance(result, dict):
-        payload = dict(result)
-        return _normalize_legacy_ocr_result(payload, default_engine=engine_label)
-    if isinstance(result, str):
-        return await _parse_text_with_legacy_parser(legacy_ocr, result, engine_label=engine_label)
-    return None
-
-
-def _is_empty_ocr_result(result: Optional[dict]) -> bool:
-    if not isinstance(result, dict):
-        return True
-    raw_text = (result.get('raw_text') or '').strip()
-    parsed = result.get('parsed_nutriments') or {}
-    has_parsed = any(value is not None for value in parsed.values()) if isinstance(parsed, dict) else False
-    return not raw_text and not has_parsed
-
-
-@_route_post(
-    "/by-barcode",
-    response_model=ProductResponse,
-    responses={
-        404: {"model": ErrorResponse, "description": "Product not found"},
-        408: {"model": ErrorResponse, "description": "Request timeout"},
-        503: {"model": ErrorResponse, "description": "Service unavailable"}
-    }
-)
-async def get_product_by_barcode(
-    request: BarcodeRequest,
-    context: RequestContext = Depends(get_optional_request_context)
-):
-    context = _ensure_request_context(context)
-    raw_barcode = request.barcode or ""
-    barcode = raw_barcode.strip()
-    cache = _get_cache_backend()
-    product_client = _get_openfoodfacts_backend()
-
-    if not barcode:
-        raise _raise_http_exception(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="Barcode cannot be empty",
-        )
-
-    if len(barcode) > 256:
-        raise _raise_http_exception(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="Barcode exceeds maximum length",
-        )
-
-    if not re.fullmatch(r"[A-Za-z0-9_\-]+", barcode):
-        raise _raise_http_exception(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="Barcode contains invalid characters",
-        )
-    start_time = datetime.now()
-    
-    user_id = context.user_id
-    session_id = context.session_id
-    timeout_exc, request_error_exc = _get_httpx_exceptions()
-    
-    try:
-        # Check cache
-        cache_key = f"product:{barcode}"
-        cached_product = await cache.get(cache_key)
-        if cached_product:
-            response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_product_lookup(
-                user_id, session_id, barcode, cached_product.get('name'), 
-                True, response_time, "Cache"
-            )
-            logger.info(f"Returning cached product for barcode: {barcode}")
-            return ProductResponse(**cached_product)
-
-        # Fetch from external API
-        try:
-            product = await product_client.get_product(barcode)
-        except timeout_exc:
-            response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_product_lookup(
-                user_id, session_id, barcode, None,
-                False, response_time, "OpenFoodFacts", "Timeout"
-            )
-            logger.error(f"Timeout fetching product for barcode: {barcode}")
-            raise _raise_http_exception(
-                status_code=status.HTTP_408_REQUEST_TIMEOUT,
-                detail="Request timeout while fetching product data"
-            )
-        except request_error_exc as e:
-            response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_product_lookup(
-                user_id, session_id, barcode, None,
-                False, response_time, "OpenFoodFacts", f"Network error: {str(e)}"
-            )
-            logger.error(f"Network error fetching product for barcode {barcode}: {e}")
-            raise _raise_http_exception(
-                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                detail="Unable to connect to product database"
-            )
-        except Exception as api_error:
-            # Log API error and convert to appropriate HTTP status
-            response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_product_lookup(
-                user_id, session_id, barcode, None,
-                False, response_time, "OpenFoodFacts", f"API error: {str(api_error)}"
-            )
-            logger.warning(f"OpenFoodFacts API error for barcode {barcode}: {api_error}")
-            raise _raise_http_exception(
-                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
-                detail="Product lookup service temporarily unavailable"
-            )
-        
-        if product is None:
-            response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_product_lookup(
-                user_id, session_id, barcode, None, 
-                False, response_time, "OpenFoodFacts", "Product not found"
-            )
-            raise _raise_http_exception(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Product with barcode {barcode} not found"
-            )
-        
-        # Store in database and cache
-        product_dict = product.model_dump()
-        await cache.set(cache_key, product_dict, ttl_hours=24)
-
-        # Store in database for future lookups using Repository Pattern
-        try:
-            product_repo = ProductRepository()
-            product_entity = Product(
-                barcode=product.barcode,
-                name=product.name,
-                brand=product.brand,
-                serving_size=product.serving_size,
-                nutriments=product_dict.get('nutriments', {})
-            )
-            await product_repo.create(product_entity)
-        except Exception as e:
-            # Log but don't fail if database storage fails - cache is sufficient
-            logger.warning(f"Failed to store product {product.barcode} in database: {e}")
-        
-        response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-        await analytics_service.log_product_lookup(
-            user_id, session_id, barcode, product.name, 
-            True, response_time, "OpenFoodFacts"
-        )
-        await analytics_service.log_user_product_interaction(
-            user_id, session_id, barcode, "lookup", "api_fetch"
-        )
-        
-        logger.info(f"Successfully fetched, stored, and cached product for barcode: {barcode}")
-        return product
-    except Exception as exc:
-        if _is_http_exception(exc):
-            raise exc
-        response_time = int((datetime.now() - start_time).total_seconds() * 1000)
-        await analytics_service.log_product_lookup(
-            user_id, session_id, barcode, None, 
-            False, response_time, "System", f"Unexpected error: {str(exc)}"
-        )
-        logger.error(f"Unexpected error fetching product for barcode {barcode}: {exc}", exc_info=True)
-        raise _raise_http_exception(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
-        )
-
-
-@_route_post(
-    "/scan-label",
-    response_model=Union[ScanResponse, LowConfidenceScanResponse],
-    responses={
-        400: {"model": ErrorResponse, "description": "Invalid image file"},
-        413: {"model": ErrorResponse, "description": "File too large"},
-        500: {"model": ErrorResponse, "description": "Processing error"}
-    }
-)
-async def scan_nutrition_label(
-    image: UploadFile = File(None, alias="image"),
-    legacy_file: UploadFile = File(None, alias="file"),
-    context: RequestContext = Depends(get_optional_request_context)
-):
-    """
-    Scan nutrition label from uploaded image using OCR
-    """
-    upload = image or legacy_file
-
-    if upload is None:
-        raise _raise_http_exception(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="field required"
-        )
-
-    # Validate file type
-    if not upload.content_type or not upload.content_type.startswith('image/'):
-        raise _raise_http_exception(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="File must be an image (JPEG, PNG, etc.)"
-        )
-    
-    # Check file size (max 10MB)
-    if upload.size and upload.size > 10 * 1024 * 1024:
-        raise _raise_http_exception(
-            status_code=status.HTTP_413_REQUEST_ENTITY_TOO_LARGE,
-            detail="Image file too large (max 10MB)"
-        )
-    
-    context = _ensure_request_context(context)
-
-    # Extract user context
-    user_id = context.user_id
-    session_id = context.session_id
-    
-    temp_file_path = None
-    start_time = datetime.now()
-    
-    try:
-        # Save uploaded file to temporary storage
-        suffix = os.path.splitext(upload.filename or '')[1] or '.jpg'
-        temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=suffix)
-        temp_file_path = temp_file.name
-        
-        # Write uploaded content to temp file
-        content = await upload.read()
-        await _write_bytes_to_tempfile(temp_file_path, content)
-        
-        logger.info(f"Image saved to temp file: {temp_file_path}")
-        
-        # Extract nutrients using enhanced OCR service
-        ocr_result = await _run_local_ocr(temp_file_path)
-        
-        if not ocr_result['raw_text'].strip():
-            processing_time_ms = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_ocr_scan(
-                user_id, session_id, upload.size, 0.0, processing_time_ms,
-                ocr_result.get('processing_details', {}).get('ocr_engine', 'tesseract'),
-                0, False, "No text extracted"
-            )
-            raise _raise_http_exception(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="No text could be extracted from the image"
-            )
-        
-        # Extract parsed data
-        nutrition_data = ocr_result['parsed_nutriments']
-        serving_size = ocr_result.get('serving_size', '100g')  # Default serving size
-        confidence = ocr_result['confidence']
-        raw_text = ocr_result['raw_text']
-        
-        processing_time = (datetime.now() - start_time).total_seconds()
-        processing_time_ms = int(processing_time * 1000)
-        logger.info(f"OCR processing completed in {processing_time:.2f}s, confidence: {confidence:.2f}")
-        
-        # Log OCR analytics
-        nutrients_extracted = len([v for v in nutrition_data.values() if v is not None])
-        await analytics_service.log_ocr_scan(
-            user_id, session_id, upload.size, confidence, processing_time_ms,
-            ocr_result.get('processing_details', {}).get('ocr_engine', 'tesseract'),
-            nutrients_extracted, True
-        )
-        
-        scan_timestamp = datetime.now()
-        
-        # High confidence response (>= 0.7)
-        if confidence >= 0.7:
-            nutriments = Nutriments(
-                energy_kcal_per_100g=nutrition_data.get('energy_kcal'),
-                protein_g_per_100g=nutrition_data.get('protein_g'),
-                fat_g_per_100g=nutrition_data.get('fat_g'),
-                carbs_g_per_100g=nutrition_data.get('carbs_g'),
-                sugars_g_per_100g=nutrition_data.get('sugars_g'),
-                salt_g_per_100g=nutrition_data.get('salt_g')
-            )
-            
-            return ScanResponse(
-                source="Local OCR",
-                confidence=confidence,
-                raw_text=raw_text,
-                serving_size=serving_size,
-                nutriments=nutriments,
-                nutrients=nutriments,
-                scanned_at=scan_timestamp
-            )
-        
-        # Low confidence response (< 0.7)
-        else:
-            return LowConfidenceScanResponse(
-                low_confidence=True,
-                confidence=confidence,
-                raw_text=raw_text,
-                partial_parsed=nutrition_data,
-                suggest_external_ocr=True,
-                scanned_at=scan_timestamp
-            )
-    
-    except Exception as exc:
-        if _is_http_exception(exc):
-            raise exc
-        processing_time_ms = int((datetime.now() - start_time).total_seconds() * 1000)
-        await analytics_service.log_ocr_scan(
-            user_id, session_id, upload.size or 0, 0.0, processing_time_ms,
-            "tesseract", 0, False, f"Processing error: {str(exc)}"
-        )
-        logger.error(f"Error processing image: {exc}")
-        raise _raise_http_exception(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Error processing image"
-        )
-    
-    finally:
-        # Clean up temp file
-        if temp_file_path:
-            try:
-                os.unlink(temp_file_path)
-                logger.info(f"Cleaned up temp file: {temp_file_path}")
-            except FileNotFoundError:
-                logger.debug(f"Temp file already removed: {temp_file_path}")
-            except OSError as e:
-                logger.warning(f"Failed to clean up temp file {temp_file_path}: {e}")
-
-
-@_route_post(
-    "/scan-label-external",
-    response_model=Union[ScanResponse, LowConfidenceScanResponse],
-    responses={
-        400: {"model": ErrorResponse, "description": "Invalid image file"},
-        503: {"model": ErrorResponse, "description": "External OCR service unavailable"}
-    }
-)
-async def scan_label_with_external_ocr(
-    image: UploadFile = File(None, alias="image"),
-    legacy_file: UploadFile = File(None, alias="file"),
-    context: RequestContext = Depends(get_optional_request_context)
-):
-    """
-    Scan nutrition label using external OCR service (when available)
-    Falls back to local OCR if external service is unavailable
-    """
-    upload = image or legacy_file
-    using_legacy_upload = legacy_file is not None and image is None
-
-    if upload is None:
-        raise _raise_http_exception(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail="field required"
-        )
-
-    # Similar validation as scan_label endpoint
-    if not upload.content_type or not upload.content_type.startswith('image/'):
-        raise _raise_http_exception(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="File must be an image"
-        )
-    
-    context = _ensure_request_context(context)
-
-    # Extract user context
-    user_id = context.user_id
-    session_id = context.session_id
-    
-    temp_file_path = None
-    start_time = datetime.now()
-    
-    try:
-        # Save uploaded file
-        suffix = os.path.splitext(upload.filename or '')[1] or '.jpg'
-        temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=suffix)
-        temp_file_path = temp_file.name
-        
-        content = await upload.read()
-        await _write_bytes_to_tempfile(temp_file_path, content)
-        
-        # Try external OCR first
-        legacy_ocr = _import_legacy_ocr()
-        external_callable = getattr(legacy_ocr, 'call_external_ocr', None) if legacy_ocr else None
-        ocr_result = None
-        if external_callable:
-            try:
-                external_payload = external_callable(temp_file_path)
-                if inspect.isawaitable(external_payload):
-                    external_payload = await external_payload
-                ocr_result = await _normalize_external_payload(external_payload, legacy_ocr, engine_label='external_ocr')
-                if _is_empty_ocr_result(ocr_result):
-                    ocr_result = None
-            except Exception as exc:
-                logger.warning(f"External OCR failed, falling back to local OCR: {exc}")
-                ocr_result = None
-
-        if ocr_result is None:
-            try:
-                fallback_payload = call_external_ocr(temp_file_path)
-                if inspect.isawaitable(fallback_payload):
-                    fallback_payload = await fallback_payload
-                ocr_result = await _normalize_external_payload(fallback_payload, legacy_ocr, engine_label='external_ocr')
-                if _is_empty_ocr_result(ocr_result):
-                    ocr_result = None
-            except Exception as exc:
-                logger.warning(f"External OCR hook failed, falling back to local OCR: {exc}")
-                ocr_result = None
-
-        if ocr_result:
-            logger.info("Using external OCR service result")
-            source = "external_ocr" if using_legacy_upload else "External OCR"
-            external_used = True
-        else:
-            logger.info("External OCR unavailable, falling back to local OCR")
-            ocr_result = await _run_local_ocr(temp_file_path)
-            source = "local_ocr_fallback" if using_legacy_upload else "Local OCR (fallback)"
-            external_used = False
-        
-        if not ocr_result['raw_text'].strip():
-            processing_time_ms = int((datetime.now() - start_time).total_seconds() * 1000)
-            await analytics_service.log_ocr_scan(
-                user_id, session_id, upload.size, 0.0, processing_time_ms,
-                source.replace(" OCR", "").lower() + "_api", 0, False, "No text extracted"
-            )
-            raise _raise_http_exception(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="No text could be extracted from the image"
-            )
-        
-        # Extract parsed data
-        nutrition_data = ocr_result['parsed_nutriments']
-        serving_size = ocr_result.get('serving_size', '100g')  # Default serving size
-        confidence = ocr_result['confidence']
-        raw_text = ocr_result['raw_text']
-        
-        processing_time = (datetime.now() - start_time).total_seconds()
-        processing_time_ms = int(processing_time * 1000)
-        logger.info(f"External OCR processing completed in {processing_time:.2f}s, confidence: {confidence:.2f}")
-        
-        # Log OCR analytics
-        nutrients_extracted = len([v for v in nutrition_data.values() if v is not None])
-        await analytics_service.log_ocr_scan(
-            user_id, session_id, upload.size, confidence, processing_time_ms,
-            ocr_result.get('processing_details', {}).get('ocr_engine', 'external_api'),
-            nutrients_extracted, True
-        )
-        
-        scan_timestamp = datetime.now()
-        
-        if confidence >= 0.7:
-            nutriments = Nutriments(
-                energy_kcal_per_100g=nutrition_data.get('energy_kcal'),
-                protein_g_per_100g=nutrition_data.get('protein_g'),
-                fat_g_per_100g=nutrition_data.get('fat_g'),
-                carbs_g_per_100g=nutrition_data.get('carbs_g'),
-                sugars_g_per_100g=nutrition_data.get('sugars_g'),
-                salt_g_per_100g=nutrition_data.get('salt_g')
-            )
-            
-            return ScanResponse(
-                source=source,
-                confidence=confidence,
-                raw_text=raw_text,
-                serving_size=serving_size,
-                nutriments=nutriments,
-                nutrients=nutriments,
-                scanned_at=scan_timestamp
-            )
-        else:
-            return LowConfidenceScanResponse(
-                low_confidence=True,
-                confidence=confidence,
-                raw_text=raw_text,
-                partial_parsed=nutrition_data,
-                suggest_external_ocr=not external_used,
-                scanned_at=scan_timestamp
-            )
-    
-    except Exception as exc:
-        if _is_http_exception(exc):
-            raise exc
-        processing_time_ms = int((datetime.now() - start_time).total_seconds() * 1000)
-        await analytics_service.log_ocr_scan(
-            user_id, session_id, upload.size or 0, 0.0, processing_time_ms,
-            "external_api", 0, False, f"Processing error: {str(exc)}"
-        )
-        logger.error(f"Error in external OCR processing: {exc}")
-        raise _raise_http_exception(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Error processing image with external OCR"
-        )
-    
-    finally:
-        if temp_file_path:
-            try:
-                os.unlink(temp_file_path)
-            except FileNotFoundError:
-                logger.debug(f"Temp file already removed: {temp_file_path}")
-            except OSError as e:
-                logger.warning(f"Failed to clean up temp file: {e}")
